# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Agzos Hub internal management system API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tools:
    get:
      operationId: listTools
      tags: [tools]
      summary: List all tools and accounts
      responses:
        "200":
          description: List of tools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tool"
    post:
      operationId: createTool
      tags: [tools]
      summary: Add a tool/account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateToolBody"
      responses:
        "201":
          description: Created tool
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tool"
  /tools/{id}:
    put:
      operationId: updateTool
      tags: [tools]
      summary: Update tool
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateToolBody"
      responses:
        "200":
          description: Updated tool
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tool"
    delete:
      operationId: deleteTool
      tags: [tools]
      summary: Delete tool
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    Tool:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [advertising, analytics, design, development, communication, finance, other]
        url:
          type: string
        loginEmail:
          type: string
        monthlyCost:
          type: number
        status:
          type: string
          enum: [active, inactive]
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, category, status, createdAt]
    CreateToolBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
          enum: [advertising, analytics, design, development, communication, finance, other]
        url:
          type: string
        loginEmail:
          type: string
        monthlyCost:
          type: number
        status:
          type: string
          enum: [active, inactive]
        notes:
          type: string
      required: [name, category, status]
    UpdateToolBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
          enum: [advertising, analytics, design, development, communication, finance, other]
        url:
          type: string
        loginEmail:
          type: string
        monthlyCost:
          type: number
        status:
          type: string
          enum: [active, inactive]
        notes:
          type: string